refactor: aligned SystematicStudyController implementation to new guidelines
